python-api/old-requirements.txt
alabaster
asn1crypto
astroid
astropy
attrs
Babel
backcall
backports.shutil-get-terminal-size
beautifulsoup4
bitarray
bkcharts
blaze
bleach
bokeh
boto
Bottleneck
certifi
cffi
chardet
click
cloudpickle
clyent
colorama
contextlib2
cryptography
cycler
Cython
cytoolz
dask
datashape
decorator
distributed
docutils
entrypoints
et-xmlfile
fastcache
filelock
Flask
Flask-Cors
gevent
glob2
gmpy2
greenlet
h5py
heapdict
html5lib
idna
imageio
imagesize
ipykernel
ipython
ipython-genutils
ipywidgets
isort
itsdangerous
jdcal
jedi
Jinja2
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab
jupyterlab-launcher
kiwisolver
lazy-object-proxy
llvmlite
locket
lxml
MarkupSafe
matplotlib
mccabe
mistune
more-itertools
mpmath
msgpack-python
multipledispatch
navigator-updater
nbconvert
nbformat
networkx
nltk
nose
notebook
numba
numexpr
numpy
numpydoc
odo
olefile
openpyxl
packaging
pandas
pandocfilters
parso
partd
path.py
pathlib2
patsy
pep8
pexpect
pickleshare
Pillow
pkginfo
pluggy
ply
prompt-toolkit
psutil
psycopg2-binary
ptyprocess
py
pycodestyle
pycosat
pycparser
pycrypto
pycurl
pyflakes
Pygments
pylint
pyodbc
pyOpenSSL
pyparsing
PySocks
pytest
pytest-arraydiff
pytest-astropy
pytest-doctestplus
pytest-openfiles
pytest-remotedata
python-dateutil
pytz
PyWavelets
PyYAML
pyzmq
QtAwesome
qtconsole
QtPy
requests
rope
ruamel-yaml
scikit-image
scikit-learn
scipy
seaborn
Send2Trash
simplegeneric
singledispatch
six
snowballstemmer
sortedcollections
sortedcontainers
Sphinx
sphinxcontrib-websupport
spyder
SQLAlchemy
statsmodels
sympy
tables
tblib
terminado
testpath
toolz
tornado
traitlets
typing
unicodecsv
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt
xlrd
XlsxWriter
xlwt
zict
